+++ lib/src/episode.dart
@@ -2,56 +2,31 @@ import 'playback_state.dart';
 
 class Episode {
   /// Episode Title
-  String _title;
+  String title;
 
   /// Episode Description
-  String _description;
+  String description;
 
   /// Episide Date
-  String _pubDate;
+  String pubDate;
 
   /// Episode File Url
-  String _url;
+  String url;
 
   /// Episode cover Url
-  String _cover;
+  String cover;
 
   /// Episode Playback Status
   PlaybackState _playback = PlaybackState.stopped;
 
   /// Constructor
   Episode(
-    var this._title,
-    var this._description,
-    var this._pubDate,
-    var this._url,
-    var this._cover,
+    this.title,
+    this.description,
+    this.pubDate,
+    this.url,
+    this.cover,
   );
-
-  /// Return Episode Title
-  String get title {
-    return _title;
-  }
-
-  /// Return Episode Description
-  String get description {
-    return _description;
-  }
-
-  /// Return Episode Date
-  String get pubDate {
-    return _pubDate;
-  }
-
-  /// Return Episode File Url
-  String get url {
-    return _url;
-  }
-
-  /// Return Episode Cover Url
-  String get cover {
-    return _cover;
-  }
 
   /// Return Episode Playback Status
   PlaybackState get playbackState {