+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
-## 1.0.0
+## 1.0.2
 
-- Podcast Class
-- Feed Class
+- Refactor
+- Fix Health Suggestions
 
 ## 1.0.1
 
 - Update README and CHANGELOG
+
+## 1.0.0
+
+- Podcast Class
+- Feed Class